[FOR-72] Copilot 리뷰 내용 검토 및 적용

+++ src/main/java/org/forif_backend/domain/staff/StaffAccountRepository.java
@@ -15,7 +15,7 @@ public interface StaffAccountRepository {
 
     void deleteById(Long userId);
 
-    List<StaffAccount> searchAdminsWithCursor(Long cursor, int size, String search);
+    List<StaffAccount> searchAdminsWithCursor(Integer cursor, int size, String search);
 
     long countAdmins(String search);
